fix(api): reuse xls style while iterating over cells
## Reasoning

While working on streaming export response, I noticed that the xls export is noticeably slower than csv export of same data. While I did not find an easy way to implement streaming xls files, I did find a way to cut response time. It appears that reusing xls styles for given format (as opposed to creating new ones for each cell) shortens the response time by over 40%.

I arrived at figure after running multiple full tox tests (for the supported versions) and comparing the medians of a response that exports ~ 3k rows. And here comes the tricky part: how can we test this time improvement reliably?

## Tests

On my machine (Retina MBP 2013), the median response time without the fix was 7.54s, and with the fix 4.3s; moreover, with the fix it was always under 5.5s. While we cannot safely hit 5.5 always, I think using 6.5 as the number to strive to is "good enough", at least for my machine. So I've added a test that tests just that.

## Summary

I'll submit a PR, and let the test run on travis to see how it fares there. Please leave a comment if you have a better way to test timing. Alternatively, if there is no need to run this timing test, I can simply remove it.